R - Flytt serializeringen inn i klassene sine
