+++ src/main/kotlin/no/nav/tiltakspenger/soknad/api/soknad/SøknadDbDTO.kt
@@ -1,7 +1,6 @@
 package no.nav.tiltakspenger.soknad.api.soknad
 
 import no.nav.tiltakspenger.soknad.api.domain.SøknadDTO
-import no.nav.tiltakspenger.soknad.api.serialize
 import no.nav.tiltakspenger.soknad.api.vedlegg.Vedlegg
 import java.time.LocalDateTime
 import java.util.UUID
@@ -20,6 +19,4 @@ data class SøknadDbDTO(
     val journalført: LocalDateTime?,
     val journalpostId: String?,
     val opprettet: LocalDateTime,
-) {
-    fun SpørsmålsbesvarelserDTO.toDbJson(): String = serialize(this)
-}
+)
